Add files via upload
This is a Database Table for an E-commerce system that shows the Fragmentation and Replication concept for the class assignment. Here, the Customer table has been copied to all locations, which shows the replication method. Order, Payment, and Product are used as regional data sources, and Customer_US/AS/EU shows the fragmentation technique.